LayoutTests/fast/dom/SelectorAPI/NSResolver-exceptions-expected.txt

This tests that NSResolver throw the correct exceptions.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS document.querySelector("xh|p", function(){}); threw exception Error: NAMESPACE_ERR: DOM Exception 14.
PASS document.querySelector("xh|p", function(){ throw "error"; }); threw exception error.
PASS document.querySelector("xh|p", {}); threw exception Error: NAMESPACE_ERR: DOM Exception 14.
PASS document.querySelector("xh|p", { lookupNamespaceURI: function(){} }); threw exception Error: NAMESPACE_ERR: DOM Exception 14.
PASS document.querySelector("xh|p", { lookupNamespaceURI: function(){ throw "error"; } }); threw exception error.
PASS document.querySelector("xh|p", { get lookupNamespaceURI() { throw "error"; } }); threw exception error.
PASS document.querySelector("xh|p", null); threw exception Error: NAMESPACE_ERR: DOM Exception 14.
PASS document.querySelector("xh|p", undefined); threw exception Error: NAMESPACE_ERR: DOM Exception 14.
PASS document.querySelector("xh|p", "A String"); threw exception Error: NAMESPACE_ERR: DOM Exception 14.
PASS successfullyParsed is true

TEST COMPLETE
